refactor(cli): split merge_request.py at the 800-code-line soft ceiling [DMD-1900]
829 code lines after the writes landed -- exactly what the RFC predicted
for eleven commands at ~75 each. CONTRIBUTING lets a file sit over the
soft ceiling until the next PR adds to it, but a brand-new module born
over it is debt on day one, so split now:

- merge_request.py      -- app, callback, the four reads; mounts the writes
- _merge_request_common -- what both need: option declarations, the ONE
                           error handler, target resolution, the
                           destructive-under-json rule, escalation, output
- _merge_request_writes -- the seven writes on their own Typer, mounted
                           flat via register(app) so permission keys stay
                           merge-request.* and --help lists one group
                           (precedent: _storage_describe.register)
- _merge_request_render -- unchanged

A third module instead of reads importing writes (or vice versa): both
import common, only merge_request imports writes -- no cycle. Behaviour
unchanged; 73 CLI tests green; every module well under 800.
